\name{dataSP}
\alias{dataSP}
\docType{data}
\title{dataSP
}
\description{
Dataset SP for usage example.
}
\usage{data("dataSP")}
\format{
  A data frame with 1000 observations on the following 4 variables.
  \describe{
    \item{\code{d}}{a numeric vector}
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{z}}{a numeric vector}
  }
}

\examples{
data(dataSP)
## maybe str(dataSP) ; plot(dataSP) ...
}
\keyword{datasets}
