\name{SecondaryPhenotype-package}
\alias{SecondaryPhenotype-package}
\alias{SecondaryPhenotype-pkg}
\docType{package}
\title{
\packageTitle{SecondaryPhenotype}
}
\description{
\packageDescription{SecondaryPhenotype}
}
\author{
\packageAuthor{SecondaryPhenotype}

Maintainer: \packageMaintainer{SecondaryPhenotype}
}
\references{
Lutz SM, Hokanson JE, Lange C. (2014) An Alternative Hypothesis Testing Strategy for Secondary Phenotype Data in Case-Control Genetic Association Studies. Frontiers in Genetics. 5 (188).
}
\keyword{ package }
